player feet stay fixed to ground until walking off a platform

+++ src/Camera.hpp
@@ -82,7 +82,7 @@ public:
 	const bool&			isOrtho() const { return ortho; }
 	const Uint32		getFramesDrawn() const { return framesDrawn; }
 	const bool&			isEnabled() const { return enabled; }
-	Uint32				getOcclusionIndex() const { return occlusionIndex; }
+	Uint64				getOcclusionIndex() const { return occlusionIndex; }
 
 	void	setClipNear(float _clipNear) { clipNear = _clipNear; }
 	void	setClipFar(float _clipFar) { clipFar = _clipFar; }
@@ -91,7 +91,7 @@ public:
 	void	setDrawMode(drawmode_t _drawMode) { drawMode = _drawMode; }
 	void	setOrtho(const bool _ortho) { ortho = _ortho; }
 	void	setEnabled(const bool _enabled) { enabled = _enabled; }
-	void	setOcclusionIndex(Uint32 _occlusionIndex) { occlusionIndex = _occlusionIndex; }
+	void	setOcclusionIndex(Uint64 _occlusionIndex) { occlusionIndex = _occlusionIndex; }
 
 	//! make a reversed-Z projection matrix with infinite range
 	//! @param radians The vertical fov
@@ -211,8 +211,8 @@ protected:
 	Renderer* renderer = nullptr;
 
 	//! occlusion data
-	Map<Uint32, Map<Entity*, occlusion_query_t>> occlusionData;
-	Uint32 occlusionIndex = 0;
+	Map<Uint64, Map<Entity*, occlusion_query_t>> occlusionData;
+	Uint64 occlusionIndex = 0;
 
 	//! drawing mode
 	drawmode_t drawMode = DRAW_STANDARD;
